+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
+### v1.1.1
+
+* Change `DragDismissBundleBuilder` to `DragDismissIntentBuilder`
+
 ### v1.1.0
 
-* Add an API to access a ProgressBar for loading data
+* Add an API to access a `ProgressBar` for loading data
 
 ### v1.0.0